move core app controller logic that will be shared with the mobile app controller to concerns
